+++ src/components/emailSelector/EmailSelector.tsx
@@ -20,6 +20,7 @@ interface EmailSelectorProps extends Omit<BoxProps, 'ref' | 'onBlur' | 'onFocus'
   onFocus?: (event: React.FocusEvent<HTMLElement>) => void;
   id?: string;
   renderSuggestion?: React.ComponentType<React.ComponentProps<typeof EmailSuggestion>>;
+  disableRemovalOfFirst?: boolean;
 }
 
 const EmailSelector = ({
@@ -33,6 +34,7 @@ const EmailSelector = ({
   suggestions,
   renderSuggestion,
   warning,
+  disableRemovalOfFirst,
   ...rest
 }: EmailSelectorProps) => {
   const ref = useRef<HTMLElement>();
@@ -68,7 +70,7 @@ const EmailSelector = ({
 
   const onLabelClick = useCallback(
     (index: number) => {
-      setEditingLabel(index);
+      !disableRemovalOfFirst && setEditingLabel(index);
       setSelection(selection.filter((selection, i) => i <= index || selection.email.trim() !== ''));
     },
     [setEditingLabel, setSelection, selection],
@@ -168,7 +170,7 @@ const EmailSelector = ({
     setEditingLabel(null);
   }, [selection, setSelection, setEditingLabel]);
 
-  const onRemove = useCallback(
+  const onRemoveHandler = useCallback(
     (index: number) => {
       if (editingLabel !== index && selection[index].email !== '') {
         onUpdateLabel(index);
@@ -212,9 +214,10 @@ const EmailSelector = ({
             onFinish={onUpdateLabel}
             onFocus={onTagFocus}
             onBlur={onBlurLabel}
-            onRemove={onRemove}
+            onRemove={onRemoveHandler}
             suggestions={validSuggestions}
             renderSuggestion={renderSuggestion}
+            disableRemovalOfFirst={disableRemovalOfFirst}
           />
         ))}
         {(editingLabel === null || selection[editingLabel].email !== '') && (